split out chunked reader
